clean up init functions
