+++ main/blink.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void blink_task();
+
