+++ main/i2c.h
@@ -1,12 +1,8 @@
 #pragma once
 // https://github.com/raspberrypi/pico-examples/tree/master/i2c
 
-#include <stddef.h>
-#include <stdint.h>
-#include <hardware/i2c.h>
-
 #define MAX_SLAVES 10
 
-/** \brief looking for slave addresses and requesting updates */
+//! looking for slave addresses and requesting updates
 void bus_task();
