+++ main/sock.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-/** \brief start listening for TCP socket requests */
+//! start listening for TCP socket requests
 void serve_task();
 
 void i2c_send(uint16_t addr, const char * data, size_t data_size);
