+++ main/tasks.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void init_tasks();
+
